django-projects/Photo/pyproject.toml
[tool.poetry]
name = "photo"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
django-ckeditor = "^6.7.2"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
django-countries = "^7.6.1"
asgiref = "^3.8.1"
pyjwt = "^2.10.1"
python-decouple = "^3.8"
pytz = "^2024.2"
sqlparse = "^0.5.2"
django-rest-passwordreset = "^1.5.0"
django-rest-knox = "^5.0.2"
django-multiselectfield = "^0.1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
